Moved folder classes into src
