+++ src/scripts/classes/index.ts
